Deployment/requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
Pillow
scikit-learn==1.3.0
